refactor(backend): hoist safe imports and late-bind services (#876)
