+++ backend/app/features/schedules/job.py
@@ -45,7 +45,7 @@ from app.features.schedules.runtime_summary import (
 from app.features.schedules.service import (
     a2a_schedule_service,
 )
-from app.integrations.a2a_client import get_a2a_service
+from app.integrations.a2a_client.service import get_a2a_service
 from app.runtime.ops_metrics import ops_metrics
 from app.runtime.ops_metrics_refresh import refresh_db_pool_checked_out
 from app.runtime.scheduler import get_scheduler
