+++ backend/app/main.py
@@ -39,7 +39,10 @@ from app.features.schedules.job import ensure_a2a_schedule_job
 from app.features.schedules.service import (
     ensure_a2a_schedule_execution_cleanup_job,
 )
-from app.integrations.a2a_client import get_a2a_service, shutdown_a2a_service
+from app.integrations.a2a_client.service import (
+    get_a2a_service,
+    shutdown_a2a_service,
+)
 from app.integrations.a2a_extensions.service import (
     get_a2a_extensions_service,
     shutdown_a2a_extensions_service,
